+++ tests/22-test.js
@@ -8,7 +8,7 @@ let tpl = IJ2TPL.parse(source);
 
 var exists = false;
 
-for (const token of tpl.treeRoot)
+for (const token of tpl.root)
 	if (token[1] === '') exists = true;
 
 expected(exists, false);